+++ src/type-extensions.ts
@@ -19,3 +19,12 @@ declare module "ethers" {
     customData: any & { txName?: string };
   }
 }
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace Truffle {
+    interface TransactionDetails {
+      txName?: string;
+    }
+  }
+}