+++ test/fixture-projects/hardhat-project-typechain-truffle/deploy/2_linking.migration.ts
@@ -20,7 +20,7 @@ export = async (deployer: Deployer) => {
     },
   });
 
-  await govToken.transferOwnership(TOKEN_OWNER);
+  await govToken.transferOwnership(TOKEN_OWNER, { txName: "Transfer Ownership" });
 
   const fundingTransaction = await deployer.sendNative(TOKEN_OWNER, 100n, "Funding Governance Token Owner #1");
   await Reporter.reportTransactionByHash(